requirements.txt
langchain
langchain_community
python-dotenv
langchain_core
streamlit
langchain-groq